Split time summary into separate component
